Trim the doc-gate tests down to the two behavioural ones
The 560-line DataGridCommitThreadingDocTests carried four source-scanning
tests -- two that asserted the XML prose itself and two name-based structural
denylists -- plus ~150 lines of Roslyn plumbing, to guard 102 lines of doc.

Both kinds are dropped:

* The prose assertions bound each threading claim to a specific <para> label,
  so any honest restructure of the doc failed the build. That is the overfit
  the alternative-solution reviewer flagged and I wrongly disputed. I had
  cited FlyoutPlacementGuardTests as precedent for prose-gating; it isn't --
  its assertions are all code-shape (No_Flyout_Placement_Write_Bypasses_
  FlyoutPlacement, Scanner_*). The repo has no prose-gating precedent.

* The structural denylists are strictly dominated by the behavioural pair.
  Mutation-proven earlier in this PR: wrapping the dispatch in a bare
  new Thread(...).Start() PASSED Dispatcher_Arm_Does_Not_Offload and was
  caught only by the behavioural test. A name denylist can only reject the
  offload primitives someone remembered to list.

What survives is the two tests that had no equivalent before this PR: they
drive the real HandleAsyncCommit and observe which thread the callback lands
on. Re-verified against four mutations, each failing only its own test:

  bare new Thread(...) around the dispatch  -> dispatcher test
  drop the dispatcher arm's early return    -> dispatcher test
  fallback offloads to a dedicated thread   -> fallback test
  fallback invokes OnRowChanged inline      -> fallback test (the #958 shape)

The last <para> of HandleAsyncCommit's remarks is reworded to match: the
tests pin what each arm does with the callback's thread, they do not gate
the wording.

12934 passed / 64 skipped, plus the known load-sensitive CompilationLoader
flake, which passes in isolation.

Copilot-Session: 34a6a85f-3ba0-4c1e-a8f6-d0a761ea5f36
